requirements.txt
-i https://pypi.python.org/simple
aiofiles==0.8.0
anyio==3.6.2 ; python_full_version >= '3.6.2'
attrs==22.2.0 ; python_version >= '3.6'
aws-cdk-lib==2.72.1
aws-cdk.asset-awscli-v1==2.2.129 ; python_version ~= '3.7'
aws-cdk.asset-kubectl-v20==2.1.1 ; python_version ~= '3.7'
aws-cdk.asset-node-proxy-agent-v5==2.0.105 ; python_version ~= '3.7'
cattrs==22.2.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
# cfn-guard - while cfn-guard must be installed, there is no pip installer for it as of April 2023
# Install guide for cfn-guard: https://docs.aws.amazon.com/cfn-guard/latest/ug/setting-up-linux.html
colorlog==4.8.0
constructs==10.1.302
exceptiongroup==1.1.1 ; python_version < '3.11'
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3
idna==3.4
importlib-resources==5.12.0 ; python_version >= '3.7'
iniconfig==2.0.0 ; python_version >= '3.7'
jsii==1.80.0 ; python_version ~= '3.7'
mergedeep==1.3.4
packaging==23.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
publication==0.0.3
py==1.11.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pytest==6.2.5
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
rfc3986[idna2008]==1.5.0
semver==2.13.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.0 ; python_version >= '3.7'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
typeguard==2.13.3 ; python_full_version >= '3.5.3'
typing-extensions==4.5.0 ; python_version >= '3.7'
zipp==3.15.0 ; python_version < '3.10'
